Struct atmega2560_hal::port::portc::PC3

pub struct PC3<MODE> { /* fields omitted */ }

Type representing a specific pin.

See the general Digital IO documentation for more info. In short:

Mode is changed with

.into_output()

.into_floating_input()

.into_pull_up_input()

Input pins are sampled with

.is_high().void_unwrap()

.is_low().void_unwrap()

Output pins are set with

.set_high().void_unwrap()

.set_low().void_unwrap()

and can be checked with

.is_set_high().void_unwrap()

.is_set_low().void_unwrap()

Pins can be downgraded into a generic type using .downgrade().

Implementations

impl<MODE: DigitalIO> PC3<MODE>

pub fn into_output<D: AsDDR>(self, ddr: &D) -> PC3<Output>

Make this pin a digital output.

pub fn into_floating_input<D: AsDDR>(self, ddr: &D) -> PC3<Input<Floating>>

Make this pin a digital input without enabling the internal pull-up.

pub fn into_pull_up_input<D: AsDDR>(self, ddr: &D) -> PC3<Input<PullUp>>

Make this pin a digital input and enable the internal pull-up.

pub fn into_tri_state<D: AsDDR>(self, ddr: &D) -> PC3<TriState>

Make this pin a tri-state pin. Default state is released (high) mode. Internal pull-up is not used.

Note that, as always, it is not safe to connect the external pull-up to a voltage higher than VCC + 0.5. See your chip's datasheet for more details.

impl<MODE> PC3<MODE>

pub fn downgrade(self) -> Pin<MODE>

Downgrade this pin into a type that is generic over all pins.

The main use for this function is to store multiple pins in an array. Please note that generic pins have a runtime overhead.

See the general Digital IO documentation and the Pin type for more information.

Example

let p1 = portb.pb1.downgrade();
let p2 = portc.pc7.downgrade();
let pins = [p1, p2];